chore: file reorganisation, separate the logic
