+++ src/main.rs
@@ -1,4 +1,5 @@
 mod cli;
+mod codeblock;
 mod command;
 mod config;
 mod runner;
